recommiting because I forgot to save this file

+++ batch_1_compilation/print_the_bigger_number.py
@@ -1,2 +1,2 @@
-first_number = float(input("Please enter a number "))
-second_number = float(input("Please enter another number "))
+first_number = int(input("Please enter a number "))
+second_number = int(input("Please enter another number "))
